Refactor test files location
