+++ src/index.ts
@@ -1,2 +1,3 @@
 export { generateUISchema } from './generateUISchema';
 export { validateJSONSchema } from './validateJsonSchema';
+import 'jest-extended';
